Assign Code dict to Message directly.
Doing this instead of assigning the code['value'], which runs risk of having a later process check for the Message's 'value' key. Since it would be a string at that point, that raises TypeError for wrong type of index (string instead of integer).

Other processes already check `if message` before getting the value of that message, so this should fix the issue downstream.